FEAT: SYNCRONIZE URLS TO SEARCH FORM STATE
There where a couple of UX issues:
1. If you click back after a search, the search form field still has the value you entered even though the list is no longer filtered
2. If you refresh the page after searching, the form no longer has the value in it, even though the list is filtered

Solutions
Returning q from the loader in root.jsx and set it as the search field default value. this will fix issue 2
Synchronize input value with the URL SEARCH PARAMS in root.jsx using the useEffect hook to manipulate the form's state in the DOM directly.