FEAT: SUBMIT FORM ONCHANGES
With the use of the useSubmit hook, as we type, the form is submitted automatically.
Note! the arguement to submit. we are passing in event.currentTarget.form.
The currentTarget.form is the DOM node the event is attached to , and the currentTarget.form is the input's parent form node. the submit function will serialize and submit any form you pass to it.